add some tests and update proposers
Currently the proposers aren't resilient at all.  For example if an
acceptor has agreed to a promise to another node, but hasn't accepted
a value.  In this case if a proposer sends a promise with a lower ballot
it will not follow up with a promise of a higher ballot.