Change accept and promise handling to be more correct
This does not include the retry mechanic in the event accepts or
promises fail

What this does include is getting a quarem of promises and taking
the highest ballot value.

Then sends accepts to all acceptors.  Does not retry if needing to
promise again.

I have no idea what happens if there are network problems and say
2/5 acceptors get a, 2/5 acceptors get b.  And c is unreachable.

This would come from A promises to 1,2,3, then accepts to 1,2
B promises to 3,4,5 then accepts to 4,5

Now 3 will either be B or whoever wins the promise war with 3, but if
3 is unreachable what happens to this round?

This also includes a bunch of rambling.  The next commit should clean up
comments and add retrying