only wait for accept messages until a value has been decided
The reason for this change is because I need to have a different way
for the proposers to be updated.

There's the learner step where they notify all proposers that I think
is why I originally was waiting for all accepts.
If i want to replicate this what I should do is send a message
to all proposers with acceptors I either didn't process or got a different
value from telling them a value has been decided and to update.